add: swap package root folder for cleaner imports
