+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 package-mode = true
 name = "mm-toolbox"
-version = "0.1.1"
+version = "0.2.0"
 description = "\"high-performance python tools for market making strategies.\""
 license = "MIT"
 authors = [
