+++ src/mm_toolbox/candles/time.py
@@ -1,4 +1,4 @@
-from mm_toolbox.src.candles.base import BaseCandles
+from src.mm_toolbox.candles.base import BaseCandles
 
 
 class TimeCandles(BaseCandles):
