+++ tests/logging/test_logging_logger.py
@@ -2,7 +2,7 @@ import asyncio
 import unittest
 from unittest.mock import patch, AsyncMock
 
-from mm_toolbox.src.logging import Logger
+from src.mm_toolbox.logging import Logger
 
 # Two commented out tests need working on, i'm finding difficultly to properly
 # ensure their function. Any help is much appreciated!
@@ -12,7 +12,7 @@ class TestLogger(unittest.IsolatedAsyncioTestCase):
     def setUp(self):
         self.logger = Logger(debug_mode=True)
 
-    # @patch("mm_toolbox.src.logging.logger.aiofiles.open", new_callable=AsyncMock)
+    # @patch("src.mm_toolbox.logging.logger.aiofiles.open", new_callable=AsyncMock)
     # async def test_write_logs_to_file(self, mock_open):
     #     self.logger.msgs = ["test log message"]
 
@@ -26,41 +26,41 @@ class TestLogger(unittest.IsolatedAsyncioTestCase):
     #     mock_file.writelines.assert_called_once_with(expected_lines)
     #     self.assertEqual(self.logger.msgs, [])
 
-    @patch("mm_toolbox.src.logging.logger.Logger._message_", new_callable=AsyncMock)
+    @patch("src.mm_toolbox.logging.logger.Logger._message_", new_callable=AsyncMock)
     async def test_critical(self, mock_message):
         await self.logger.critical("test_topic", "test_message")
         mock_message.assert_called_once_with(
             "CRITICAL", "TEST_TOPIC", "test_message", flush_buffer=True
         )
 
-    @patch("mm_toolbox.src.logging.logger.Logger._message_", new_callable=AsyncMock)
+    @patch("src.mm_toolbox.logging.logger.Logger._message_", new_callable=AsyncMock)
     async def test_debug(self, mock_message):
         await self.logger.debug("test_topic", "test_message")
         mock_message.assert_called_once_with("DEBUG", "TEST_TOPIC", "test_message")
 
-    @patch("mm_toolbox.src.logging.logger.Logger._message_", new_callable=AsyncMock)
+    @patch("src.mm_toolbox.logging.logger.Logger._message_", new_callable=AsyncMock)
     async def test_error(self, mock_message):
         await self.logger.error("test_topic", "test_message")
         mock_message.assert_called_once_with("ERROR", "TEST_TOPIC", "test_message")
 
-    @patch("mm_toolbox.src.logging.logger.Logger._message_", new_callable=AsyncMock)
+    @patch("src.mm_toolbox.logging.logger.Logger._message_", new_callable=AsyncMock)
     async def test_info(self, mock_message):
         await self.logger.info("test_topic", "test_message")
         mock_message.assert_called_once_with("INFO", "TEST_TOPIC", "test_message")
 
-    @patch("mm_toolbox.src.logging.logger.Logger._message_", new_callable=AsyncMock)
+    @patch("src.mm_toolbox.logging.logger.Logger._message_", new_callable=AsyncMock)
     async def test_success(self, mock_message):
         await self.logger.success("test_topic", "test_message")
         mock_message.assert_called_once_with("SUCCESS", "TEST_TOPIC", "test_message")
 
-    @patch("mm_toolbox.src.logging.logger.Logger._message_", new_callable=AsyncMock)
+    @patch("src.mm_toolbox.logging.logger.Logger._message_", new_callable=AsyncMock)
     async def test_warning(self, mock_message):
         await self.logger.warning("test_topic", "test_message")
         mock_message.assert_called_once_with("WARNING", "TEST_TOPIC", "test_message")
 
-    # @patch("mm_toolbox.src.logging.logger.DiscordClient", autospec=True)
-    # @patch("mm_toolbox.src.logging.logger.TelegramClient", autospec=True)
-    # @patch("mm_toolbox.src.logging.logger.time_iso8601", return_value="2024-08-23T12:00:00Z")
+    # @patch("src.mm_toolbox.logging.logger.DiscordClient", autospec=True)
+    # @patch("src.mm_toolbox.logging.logger.TelegramClient", autospec=True)
+    # @patch("src.mm_toolbox.logging.logger.time_iso8601", return_value="2024-08-23T12:00:00Z")
     # async def test_message(self, mock_time_iso8601, MockTelegramClient, MockDiscordClient):
     #     self.logger.discord_client = MockDiscordClient()
     #     self.logger.telegram_client = MockTelegramClient()
@@ -73,8 +73,8 @@ class TestLogger(unittest.IsolatedAsyncioTestCase):
     #     self.logger.telegram_client.send.assert_called_once_with(expected_message, False)
     #     self.assertIn(expected_message, self.logger.msgs)
 
-    @patch("mm_toolbox.src.logging.logger.DiscordClient", autospec=True)
-    @patch("mm_toolbox.src.logging.logger.TelegramClient", autospec=True)
+    @patch("src.mm_toolbox.logging.logger.DiscordClient", autospec=True)
+    @patch("src.mm_toolbox.logging.logger.TelegramClient", autospec=True)
     async def test_shutdown(self, MockTelegramClient, MockDiscordClient):
         self.logger.discord_client = MockDiscordClient()
         self.logger.telegram_client = MockTelegramClient()
