+++ tests/numba/test_numba_linalg.py
@@ -1,6 +1,6 @@
 import unittest
 import numpy as np
-from mm_toolbox.src.numba import (
+from src.mm_toolbox.numba import (
     nbdot,
     nbnorm,
     nbsolve,
