+++ tests/orderbook/test_orderbook_hft.py
@@ -15,7 +15,7 @@
 # import unittest
 # import numpy as np
 
-# from mm_toolbox.src.orderbook import HFTOrderbook
+# from src.mm_toolbox.orderbook import HFTOrderbook
 
 # class TestHFTOrderbook(unittest.TestCase):
 #     def setUp(self):
